chore: extract common orderbook types
