+++ crates/orderbook-commons/Cargo.toml
@@ -7,9 +7,11 @@ edition = "2021"
 
 [dependencies]
 anyhow = "1"
+rust_decimal = { version = "1", features = ["serde-with-float"] }
 secp256k1 = { version = "0.24.3", features = ["serde"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 sha2 = { version = "0.10", default-features = false }
 time = { version = "0.3.20", features = ["serde"] }
+trade = { path = "../trade" }
 uuid = { version = "1.3.0", features = ["serde"] }
